requirements.txt
Django==5.1.5
djangorestframework
psycopg2-binary
